+++ database/migrations/2024_06_05_150634_create_pesanan_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users');
             $table->foreignId('produk_id')->constrained('produk');
             $table->float('jumlah');
-            $table->float('total');
+            $table->int('total');
             $table->string('alamat');
             $table->enum('status', ['Menunggu Konfirmasi', 'Sedang Dikirim', 'Selesai', 'Dibatalkan'])->default('Menunggu Konfirmasi');
             $table->mediumText('catatan')->nullable();
